Moves all scripts except the main entry to 'src' subfolder
