+++ src/animated_sprite.py
