+++ src/animation.py
