+++ src/image_resources.py
