+++ src/particle.py
